Menu Images
The menu images are in place and almost ready to go, except for the
child hover, still figuring that bit out. Also modified header.php to
include search form and fix menu bug that caused the menu to collapse
upon adding a new page. And the slider background is standard CSS now.